+++ src/applications/appeals/10182/tests/validations/issues.unit.spec.js
@@ -3,46 +3,7 @@ import sinon from 'sinon';
 
 import { maxNameLength } from '../../validations/issues';
 
-import { MAX_LENGTH, SELECTED } from '../../../shared/constants';
-import { selectionRequired } from '../../../shared/validations/issues';
-
-const _ = null;
-
-describe('selectionRequired', () => {
-  const getData = (selectContested = false, selectAdditional = false) => ({
-    contestedIssues: [
-      {
-        attributes: {
-          ratingIssueSubjectText: 'test',
-          approxDecisionDate: '2021-01-01',
-        },
-        [SELECTED]: selectContested,
-      },
-    ],
-    additionalIssues: [
-      {
-        issue: 'test 2',
-        decisionDate: '2021-01-01',
-        [SELECTED]: selectAdditional,
-      },
-    ],
-  });
-  it('should show an error when no issues are selected', () => {
-    const errors = { addError: sinon.spy() };
-    selectionRequired(errors, _, getData());
-    expect(errors.addError.called).to.be.true;
-  });
-  it('should show not show an error when a contestable issue is selected', () => {
-    const errors = { addError: sinon.spy() };
-    selectionRequired(errors, _, getData(true));
-    expect(errors.addError.called).to.be.false;
-  });
-  it('should show not show an error when an additional issue is selected', () => {
-    const errors = { addError: sinon.spy() };
-    selectionRequired(errors, _, getData(false, true));
-    expect(errors.addError.called).to.be.false;
-  });
-});
+import { MAX_LENGTH } from '../../../shared/constants';
 
 describe('maxNameLength', () => {
   it('should show an error when a name is too long', () => {
