+++ src/applications/appeals/995/tests/validations/issues.unit.spec.js
@@ -4,7 +4,6 @@ import sinon from 'sinon';
 import { checkIssues, maxNameLength } from '../../validations/issues';
 
 import { MAX_LENGTH, SELECTED } from '../../../shared/constants';
-import { selectionRequired } from '../../../shared/validations/issues';
 
 describe('checkIssues', () => {
   const _ = null;
@@ -116,43 +115,6 @@ describe('checkIssues', () => {
     const errors = { addError: sinon.spy() };
     checkIssues(errors, _, _, _, _, _, getData({ aiDate: '2021-?-?' }));
     expect(errors.addError.called).to.be.true;
-  });
-});
-
-describe('selectionRequired', () => {
-  const _ = null;
-  const getData = (selectContested = false, selectAdditional = false) => ({
-    contestedIssues: [
-      {
-        attributes: {
-          ratingIssueSubjectText: 'test',
-          approxDecisionDate: '2021-01-01',
-        },
-        [SELECTED]: selectContested,
-      },
-    ],
-    additionalIssues: [
-      {
-        issue: 'test 2',
-        decisionDate: '2021-01-01',
-        [SELECTED]: selectAdditional,
-      },
-    ],
-  });
-  it('should show an error when no issues are selected', () => {
-    const errors = { addError: sinon.spy() };
-    selectionRequired(errors, _, getData());
-    expect(errors.addError.called).to.be.true;
-  });
-  it('should show not show an error when a contested issue is selected', () => {
-    const errors = { addError: sinon.spy() };
-    selectionRequired(errors, _, getData(true));
-    expect(errors.addError.called).to.be.false;
-  });
-  it('should show not show an error when an additional issue is selected', () => {
-    const errors = { addError: sinon.spy() };
-    selectionRequired(errors, _, getData(false, true));
-    expect(errors.addError.called).to.be.false;
   });
 });
